// src/assets/types/Link.ts
/* tslint:disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

/**
 * A named resource link of a particular type.
 */
export interface Link {
  /**
   * A few words describing where this link points.
   */
  title: string;
  /**
   * An absolute URL of the resource.
   */
  href: string;
  /**
   * Link type (optional, value depending on the context).
   */
  type?: string;
}